+++ code-review-section.el
@@ -262,6 +262,8 @@ For internal usage only.")
 (defvar code-review-feedback-section-map
   (let ((map (make-sparse-keymap)))
     (define-key map (kbd "RET") 'code-review-comment-set-feedback)
+    (define-key map (kbd "C-c C-k") 'code-review-comment-delete-feedback)
+    (define-key map (kbd "k") 'code-review-comment-delete-feedback)
     map)
   "Keymaps for feedback section.")
 
@@ -567,6 +569,7 @@ Optionally DELETE? flag must be set if you want to remove it."
   (let ((map (make-sparse-keymap)))
     (define-key map (kbd "RET") 'code-review-comment-add-or-edit)
     (define-key map (kbd "C-c C-k") 'code-review-section-delete-comment)
+    (define-key map (kbd "k") 'code-review-section-delete-comment)
     map)
   "Keymaps for local-comment sections.")
 
@@ -574,6 +577,7 @@ Optionally DELETE? flag must be set if you want to remove it."
   (let ((map (make-sparse-keymap)))
     (define-key map (kbd "RET") 'code-review-comment-add-or-edit)
     (define-key map (kbd "C-c C-k") 'code-review-section-delete-comment)
+    (define-key map (kbd "k") 'code-review-section-delete-comment)
     map)
   "Keymaps for reply-comment sections.")